Add flushing of batches based on age
Add the ability to flush writes to the database if they hang out in memory for too long. This limits the damage of a crashed server and improves the recency of the data in some cases.

Also, when flushing batches based on size, the "size" is the number of pending addresses to be written in the batch, and not the number of application-level writes. This does a better job of bounding the memory usage.