Started implementing/refactoring to allow batching for slices
Ended up splitting up functionality of 'Op's into multiple
interfaces, most notably introducing IRowOp (to be renamed to
fit projects naming convention - was lacking a name for the
implementing class so far), which holds multiple serializable operations
for single columns. This allows to reduce the "original" case to an
operation on a column "with a special name" (empty string, this is
btw a TODO, since get() in harnesses accesses a field from datacube).

So far the modifications, imho, make handling key-value stored with
support for columns or hashmaps as values more generic and in the case
of increment in hbase more performant/atomic. CAS on the other hand
becomes a little more fragile due to higher chance of
conflicts/inconsistencies when updating multiple columns, since
every read-combine-cas cycle is executed independently for every column.

Those changes break the backfill functionality - TBD.