Update many dependencies and increase minor version, 1.2.x -> 1.3.0
This will be a breaking dependency change for anyone using older
versions of the following transitive dependencies.

* Guava 12.0.1 -> 15.9
  Note, this update required added NullOutputStream to the test suite to fix a runtime dependency in the hbase minicluster.
* Commons Codec 1.6 -> 1.9
* Joda Time 1.6.2 -> 2.3
* Junit 4.5 -> 4.11

We will maintain the yammer metrics version at 2.1.2 for now, but there
is a need to start planning a move to 3.0 soon. The new metrics library
version uses a different package path so there should be little
conflict. This will, of course, be tested further.